Forward CSP, even for the initial empty document.
Bug 1064676 has been fixed by:
  https://chromium-review.googlesource.com/c/chromium/src/+/2111170
And tested by:
  https://chromium-review.googlesource.com/c/chromium/src/+/2144012

The bug was fixed for every CSP checked in the renderer process. However
there are still an issue for the one checked in the browser process. It
turns out the CSP in the initial empty document weren't properly
propagated to the browser process.

This patch:
  1) Fix the bug by sending the CSP of the initial empty document.
  2) Add a regression test (WPT).

This patch can potentially also fix:
 - https://crbug.com/1072719
 - https://crbug.com/955350
(I haven't checked. I will do it later after landing this patch)

Bug: 1064676, 1072719, 955350
Change-Id: Ie5325035c74d9e2476d6c80af3e5d5c9068ea928