bluetooth: Split bluetooth-helpers.js
This change splits the bluetooth-helpers.js files in two. The new
bluetooth-test.js file contains methods for setting up the web tests and
for asserting conditions. The bluetooth-helpers.js file now contains
helper methods that set up fake Bluetooth devices for use in the Web
Bluetooth tests.

I added JsDoc style comments to the helper methods to make it clear
what kinds of objects are expected as parameters and what the methods
will return. I also reordered the bluetooth-helpers.js methods so that
they are defined before they are used.

Bug: 1070816
Change-Id: Ib5215555be2b24b11f8dea9098d83ad69719a284